Revert "Test coverage"
