+++ documentation/project_conventions/code_review_guide.md
@@ -47,7 +47,6 @@ fix(router): fix payload parameter in post request // header
   - Naming consistency and the StandardJS style guide
   - Basic security (input validation)
   - Acceptance test presence
-  - Test Coverage (no less than 80% via coveralls report)
 - Reviews must be meaningful and include **substantive direct code comments** explaining any issues.
 - ⁠Approve and merge only if code is completely satisfactory and issues are corrected.
