+++ package.json
@@ -13,7 +13,6 @@
   "standard": {
     "globals": [
       "afterAll",
-      "afterEach",
       "beforeAll",
       "beforeEach",
       "describe",
