+++ src/integration/run_integration_checks.js
@@ -1,14 +1,12 @@
 const { connectToDatabase, closeDatabaseConnection } = require('../models/db')
 const { setLecturerAvailability, getLecturerAvailability } = require('../models/lecturer_availability_db')
 
-const DEFAULT_TEST_USER = 'integration_test_lecturer'
-
 const run = async function () {
   try {
     await connectToDatabase()
     console.log('Connected to database')
 
-    const testUser = process.env.INTEG_TEST_USER || DEFAULT_TEST_USER
+    const testUser = process.env.INTEG_TEST_USER || 'integration_test_lecturer'
     const prefs = {
       minStudents: 1,
       maxStudents: 3,
@@ -28,25 +26,17 @@ const run = async function () {
 
     if (fetched && fetched.minStudents === prefs.minStudents) {
       console.log('Integration check passed')
-      return 0
+      process.exit(0)
     }
 
     console.error('Integration check failed: fetched data mismatch')
-    return 2
+    process.exit(2)
   } catch (err) {
     console.error('Integration check error:', err && err.message ? err.message : err)
-    return 1
+    process.exit(1)
   } finally {
     try { await closeDatabaseConnection() } catch (_) {}
   }
 }
 
-if (require.main === module) {
-  run().then(function (exitCode) {
-    process.exit(exitCode)
-  })
-}
-
-module.exports = {
-  run
-}
+run()
