--- tests/app.test.js
@@ -1,40 +1,0 @@
-const http = require('node:http')
-
-const app = require('../src/app')
-
-describe('application entry routes', () => {
-  let server
-  let baseUrl
-
-  beforeAll(async () => {
-    server = http.createServer(app)
-    await new Promise((resolve) => {
-      server.listen(0, '127.0.0.1', () => {
-        baseUrl = `http://127.0.0.1:${server.address().port}`
-        resolve()
-      })
-    })
-  })
-
-  afterAll(async () => {
-    await new Promise((resolve, reject) => {
-      server.close((error) => {
-        if (error) {
-          reject(error)
-          return
-        }
-
-        resolve()
-      })
-    })
-  })
-
-  test('redirects the root route to login', async () => {
-    const response = await fetch(baseUrl, {
-      redirect: 'manual'
-    })
-
-    expect(response.status).toBe(302)
-    expect(response.headers.get('location')).toBe('/login')
-  })
-})
